Code cleanup

+++ src/main/java/de/unistuttgart/iste/gits/skilllevel_service/service/calculation/SkillLevelCalculator.java
@@ -163,7 +163,7 @@ public class SkillLevelCalculator {
                 } else {
                     // learning interval is doubled for each repetition
                     Duration minimumLearningInterval = Duration.ofDays(
-                            (long)(assessment.getAssessmentMetadata().getInitialLearningInterval() * Math.pow(2, countedRepetitions - 1))
+                            (long)(assessment.getAssessmentMetadata().getInitialLearningInterval() * Math.pow(2, (int)(countedRepetitions - 1)))
                     );
                     // if the time gap between the current log item and the previous one is large enough, count it
                     // as a repetition
